HWW scale factors for leptons
