+++ data/ScaleFactors/Electrons_HWW_CutBasedID_TightWP_fromTemplates_withSyst_v1_reco_id_iso.json
@@ -1,0 +1,1 @@
+{"error_type": "absolute", "data": [{"bin": [-2.5, -2.0], "values": [{"bin": [10.0, 20.0], "error_low": 0.6290373563665135, "value": 1.2517482517482519, "error_high": 0.6290373563665135}, {"bin": [20.0, 30.0], "error_low": 0.07175182509763837, "value": 1.0585774058577406, "error_high": 0.07175182509763837}, {"bin": [30.0, 40.0], "error_low": 0.04320451866042935, "value": 1.0245901639344261, "error_high": 0.04320451866042935}, {"bin": [40.0, 50.0], "error_low": 0.023282048830689538, "value": 1.0101302460202606, "error_high": 0.023282048830689538}, {"bin": [50.0, 200.0], "error_low": 0.034123851667539146, "value": 0.9986522911051213, "error_high": 0.034123851667539146}]}, {"bin": [-2.0, -1.566], "values": [{"bin": [10.0, 20.0], "error_low": 0.4385795837633762, "value": 1.138888888888889, "error_high": 0.4385795837633762}, {"bin": [20.0, 30.0], "error_low": 0.05614243700286303, "value": 0.9838056680161943, "error_high": 0.05614243700286303}, {"bin": [30.0, 40.0], "error_low": 0.014692479113830124, "value": 0.9686028257456829, "error_high": 0.014692479113830124}, {"bin": [40.0, 50.0], "error_low": 0.0285765034381969, "value": 0.9819444444444444, "error_high": 0.0285765034381969}, {"bin": [50.0, 200.0], "error_low": 0.042073539242370346, "value": 0.9805699481865284, "error_high": 0.042073539242370346}]}, {"bin": [-1.566, -1.444], "values": [{"bin": [10.0, 20.0], "error_low": 3.4152249322509456, "value": 1.3571428571428572, "error_high": 3.4152249322509456}, {"bin": [20.0, 30.0], "error_low": 1.1928557391307972, "value": 1.186206896551724, "error_high": 1.1928557391307972}, {"bin": [30.0, 40.0], "error_low": 0.0551813663606372, "value": 1.0151515151515151, "error_high": 0.0551813663606372}, {"bin": [40.0, 50.0], "error_low": 0.02796188850166296, "value": 0.9934102141680395, "error_high": 0.02796188850166296}, {"bin": [50.0, 200.0], "error_low": 0.0957581488392178, "value": 0.9448698315467075, "error_high": 0.0957581488392178}]}, {"bin": [-1.444, -0.8], "values": [{"bin": [10.0, 20.0], "error_low": 0.8973653986277699, "value": 1.3102310231023104, "error_high": 0.8973653986277699}, {"bin": [20.0, 30.0], "error_low": 0.1299131154556001, "value": 1.0502008032128514, "error_high": 0.1299131154556001}, {"bin": [30.0, 40.0], "error_low": 0.022134301509655594, "value": 0.9891808346213292, "error_high": 0.022134301509655594}, {"bin": [40.0, 50.0], "error_low": 0.01151883524724252, "value": 0.9703903095558546, "error_high": 0.01151883524724252}, {"bin": [50.0, 200.0], "error_low": 0.03490707235646965, "value": 0.9562499999999999, "error_high": 0.03490707235646965}]}, {"bin": [-0.8, 0.0], "values": [{"bin": [10.0, 20.0], "error_low": 0.6375152908063862, "value": 1.2035398230088494, "error_high": 0.6375152908063862}, {"bin": [20.0, 30.0], "error_low": 0.11172250267410505, "value": 1.0188323917137476, "error_high": 0.11172250267410505}, {"bin": [30.0, 40.0], "error_low": 0.02583907178590443, "value": 0.9643916913946587, "error_high": 0.02583907178590443}, {"bin": [40.0, 50.0], "error_low": 0.01750660656101116, "value": 0.9564643799472295, "error_high": 0.01750660656101116}, {"bin": [50.0, 200.0], "error_low": 0.043321712841870086, "value": 0.9543773119605425, "error_high": 0.043321712841870086}]}, {"bin": [0.0, 0.8], "values": [{"bin": [10.0, 20.0], "error_low": 0.38722697347981416, "value": 1.183139534883721, "error_high": 0.38722697347981416}, {"bin": [20.0, 30.0], "error_low": 0.03775986803068175, "value": 0.994392523364486, "error_high": 0.03775986803068175}, {"bin": [30.0, 40.0], "error_low": 0.03105354442743249, "value": 0.9673590504451038, "error_high": 0.03105354442743249}, {"bin": [40.0, 50.0], "error_low": 0.01710370870535496, "value": 0.9617414248021108, "error_high": 0.01710370870535496}, {"bin": [50.0, 200.0], "error_low": 0.05652547960884774, "value": 0.9617283950617284, "error_high": 0.05652547960884774}]}, {"bin": [0.8, 1.444], "values": [{"bin": [10.0, 20.0], "error_low": 0.7156078662427867, "value": 1.3120805369127517, "error_high": 0.7156078662427867}, {"bin": [20.0, 30.0], "error_low": 0.1581784573238083, "value": 1.05668016194332, "error_high": 0.1581784573238083}, {"bin": [30.0, 40.0], "error_low": 0.019519872583171824, "value": 0.9968944099378882, "error_high": 0.019519872583171824}, {"bin": [40.0, 50.0], "error_low": 0.008884588786200623, "value": 0.9729729729729729, "error_high": 0.008884588786200623}, {"bin": [50.0, 200.0], "error_low": 0.024741747021238263, "value": 0.9560853199498117, "error_high": 0.024741747021238263}]}, {"bin": [1.444, 1.566], "values": [{"bin": [10.0, 20.0], "error_low": 3.8488051456306027, "value": 1.4878048780487805, "error_high": 3.8488051456306027}, {"bin": [20.0, 30.0], "error_low": 1.0677235095001552, "value": 1.0600706713780919, "error_high": 1.0677235095001552}, {"bin": [30.0, 40.0], "error_low": 0.18246490980540825, "value": 0.9868131868131867, "error_high": 0.18246490980540825}, {"bin": [40.0, 50.0], "error_low": 0.04858124865775569, "value": 0.9568106312292358, "error_high": 0.04858124865775569}, {"bin": [50.0, 200.0], "error_low": 0.12860988441879223, "value": 0.9493087557603687, "error_high": 0.12860988441879223}]}, {"bin": [1.566, 2.0], "values": [{"bin": [10.0, 20.0], "error_low": 0.33066100993942094, "value": 1.0845070422535212, "error_high": 0.33066100993942094}, {"bin": [20.0, 30.0], "error_low": 0.11979260581175033, "value": 1.0225409836065573, "error_high": 0.11979260581175033}, {"bin": [30.0, 40.0], "error_low": 0.01403033225696814, "value": 0.9669811320754716, "error_high": 0.01403033225696814}, {"bin": [40.0, 50.0], "error_low": 0.020613723673730752, "value": 0.9847222222222222, "error_high": 0.020613723673730752}, {"bin": [50.0, 200.0], "error_low": 0.008822867658223732, "value": 0.9922279792746114, "error_high": 0.008822867658223732}]}, {"bin": [2.0, 2.5], "values": [{"bin": [10.0, 20.0], "error_low": 0.4566325912237631, "value": 1.174061433447099, "error_high": 0.4566325912237631}, {"bin": [20.0, 30.0], "error_low": 0.06947630925972713, "value": 1.0560165975103735, "error_high": 0.06947630925972713}, {"bin": [30.0, 40.0], "error_low": 0.06988661672282794, "value": 1.022875816993464, "error_high": 0.06988661672282794}, {"bin": [40.0, 50.0], "error_low": 0.028613251205540365, "value": 1.004341534008683, "error_high": 0.028613251205540365}, {"bin": [50.0, 200.0], "error_low": 0.02032834113173589, "value": 1.001349527665317, "error_high": 0.02032834113173589}]}], "dimension": 2, "binning": {"y": [10.0, 20.0, 30.0, 40.0, 50.0, 200.0], "x": [-2.5, -2.0, -1.566, -1.444, -0.8, 0.0, 0.8, 1.444, 1.566, 2.0, 2.5]}}
