+++ data/ScaleFactors/sources/electrons/HWW_CutBasedID_TightWP_fromTemplates_withSyst_v1.txt
@@ -1,0 +1,51 @@
+-2.500  -2.000  10.000  20.000  0.358   0.009   0.286   0.002   0.094   0.048   0.071   0.127   -1  -1
+-2.000  -1.566  10.000  20.000  0.328   0.009   0.288   0.003   0.064   0.059   0.005   0.091   -1  -1
+-1.566  -1.444  10.000  20.000  0.228   0.019   0.168   0.004   0.039   0.005   0.016   0.572   -1  -1
+-1.444  -0.800  10.000  20.000  0.397   0.007   0.303   0.002   0.052   0.036   0.027   0.263   -1  -1
+-0.800  0.000   10.000  20.000  0.408   0.006   0.339   0.002   0.066   0.088   0.033   0.183   -1  -1
+0.000   0.800   10.000  20.000  0.407   0.005   0.344   0.002   0.036   0.045   0.001   0.120   -1  -1
+0.800   1.444   10.000  20.000  0.391   0.009   0.298   0.002   0.060   0.066   0.037   0.190   -1  -1
+1.444   1.566   10.000  20.000  0.244   0.014   0.164   0.004   0.043   0.020   0.041   0.628   -1  -1
+1.566   2.000   10.000  20.000  0.308   0.007   0.284   0.002   0.020   0.004   0.031   0.086   -1  -1
+2.000   2.500   10.000  20.000  0.344   0.017   0.293   0.003   0.096   0.061   0.067   0.016   -1  -1
+-2.500  -2.000  20.000  30.000  0.506   0.005   0.478   0.002   0.008   0.005   0.010   0.031   -1  -1
+-2.000  -1.566  20.000  30.000  0.486   0.006   0.494   0.002   0.008   0.012   0.018   0.014   -1  -1
+-1.566  -1.444  20.000  30.000  0.344   0.014   0.290   0.003   0.066   0.132   0.077   0.303   -1  -1
+-1.444  -0.800  20.000  30.000  0.523   0.005   0.498   0.002   0.003   0.010   0.009   0.063   -1  -1
+-0.800  0.000   20.000  30.000  0.541   0.004   0.531   0.001   0.015   0.027   0.007   0.050   -1  -1
+0.000   0.800   20.000  30.000  0.532   0.003   0.535   0.001   0.007   0.006   0.012   0.013   -1  -1
+0.800   1.444   20.000  30.000  0.522   0.007   0.494   0.002   0.002   0.023   0.014   0.073   -1  -1
+1.444   1.566   20.000  30.000  0.300   0.012   0.283   0.003   0.227   0.046   0.091   0.171   -1  -1
+1.566   2.000   20.000  30.000  0.499   0.006   0.488   0.002   0.027   0.043   0.015   0.024   -1  -1
+2.000   2.500   20.000  30.000  0.509   0.005   0.482   0.002   0.007   0.011   0.005   0.030   -1  -1
+-2.500  -2.000  30.000  40.000  0.625   0.003   0.610   0.002   0.007   0.002   0.002   0.025   -1  -1
+-2.000  -1.566  30.000  40.000  0.617   0.003   0.637   0.002   0.004   0.003   0.007   0.000   -1  -1
+-1.566  -1.444  30.000  40.000  0.469   0.006   0.462   0.003   0.002   0.005   0.001   0.024   -1  -1
+-1.444  -0.800  30.000  40.000  0.640   0.002   0.647   0.001   0.000   0.000   0.002   0.014   -1  -1
+-0.800  0.000   30.000  40.000  0.650   0.002   0.674   0.001   0.016   0.004   0.001   0.005   -1  -1
+0.000   0.800   30.000  40.000  0.652   0.001   0.674   0.001   0.020   0.002   0.004   0.004   -1  -1
+0.800   1.444   30.000  40.000  0.642   0.002   0.644   0.001   0.001   0.002   0.002   0.012   -1  -1
+1.444   1.566   30.000  40.000  0.449   0.053   0.455   0.003   0.005   0.009   0.023   0.058   -1  -1
+1.566   2.000   30.000  40.000  0.615   0.003   0.636   0.002   0.005   0.002   0.001   0.006   -1  -1
+2.000   2.500   30.000  40.000  0.626   0.028   0.612   0.002   0.012   0.012   0.002   0.028   -1  -1
+-2.500  -2.000  40.000  50.000  0.698   0.003   0.691   0.002   0.012   0.001   0.001   0.010   -1  -1
+-2.000  -1.566  40.000  50.000  0.707   0.003   0.720   0.002   0.011   0.000   0.000   0.017   -1  -1
+-1.566  -1.444  40.000  50.000  0.603   0.009   0.607   0.003   0.004   0.006   0.001   0.012   -1  -1
+-1.444  -0.800  40.000  50.000  0.721   0.002   0.743   0.001   0.004   0.004   0.000   0.006   -1  -1
+-0.800  0.000   40.000  50.000  0.725   0.001   0.758   0.001   0.002   0.005   0.001   0.012   -1  -1
+0.000   0.800   40.000  50.000  0.729   0.001   0.758   0.001   0.003   0.006   0.000   0.011   -1  -1
+0.800   1.444   40.000  50.000  0.720   0.002   0.740   0.001   0.000   0.003   0.002   0.005   -1  -1
+1.444   1.566   40.000  50.000  0.576   0.006   0.602   0.003   0.018   0.011   0.001   0.019   -1  -1
+1.566   2.000   40.000  50.000  0.709   0.003   0.720   0.002   0.013   0.001   0.001   0.006   -1  -1
+2.000   2.500   40.000  50.000  0.694   0.003   0.691   0.002   0.014   0.001   0.010   0.009   -1  -1
+-2.500  -2.000  50.000  200.000 0.741   0.006   0.742   0.004   0.024   0.000   0.002   0.003   -1  -1
+-2.000  -1.566  50.000  200.000 0.757   0.005   0.772   0.003   0.025   0.003   0.005   0.019   -1  -1
+-1.566  -1.444  50.000  200.000 0.617   0.010   0.653   0.006   0.026   0.015   0.019   0.050   -1  -1
+-1.444  -0.800  50.000  200.000 0.765   0.003   0.800   0.002   0.007   0.005   0.004   0.026   -1  -1
+-0.800  0.000   50.000  200.000 0.774   0.002   0.811   0.002   0.030   0.001   0.001   0.018   -1  -1
+0.000   0.800   50.000  200.000 0.779   0.002   0.810   0.002   0.040   0.000   0.002   0.022   -1  -1
+0.800   1.444   50.000  200.000 0.762   0.003   0.797   0.002   0.009   0.001   0.002   0.017   -1  -1
+1.444   1.566   50.000  200.000 0.618   0.009   0.651   0.006   0.004   0.028   0.000   0.078   -1  -1
+1.566   2.000   50.000  200.000 0.766   0.005   0.772   0.003   0.001   0.003   0.001   0.001   -1  -1
+2.000   2.500   50.000  200.000 0.742   0.006   0.741   0.004   0.001   0.001   0.002   0.013   -1  -1
+2.000   2.500   50.000  200.000 0.742   0.006   0.741   0.004   0.001   0.001   0.002   0.013   -1  -1
