+++ data/ScaleFactors/sources/electrons/README
@@ -2,3 +2,7 @@ Files found here:
  - /afs/cern.ch/work/a/arun/public/forEGM/CutBasedId_TextFiles/
 
 Path found on slides: https://indico.cern.ch/event/467663/contribution/4/attachments/1201335/1748331/CutBased_Id_SF_8Dec_v1.pdf
+
+For HWW Scale-Factors:
+ - https://twiki.cern.ch/twiki/pub/CMS/HWW2015TriggerResults/CutBasedID_TightWP_fromTemplates_withSyst_v1.txt
+ - https://twiki.cern.ch/twiki/bin/view/CMS/HWW2015TriggerResults
