+++ python/MuonsProducer.py
@@ -19,6 +19,8 @@ default_configuration = cms.PSet(
                 iso_loose_id_tight = cms.untracked.FileInPath('cp3_llbb/Framework/data/ScaleFactors/Muon_LooseRelIso_TightID_iso.json'),
                 iso_tight_id_medium = cms.untracked.FileInPath('cp3_llbb/Framework/data/ScaleFactors/Muon_TightRelIso_MediumID_iso.json'),
                 iso_tight_id_tight = cms.untracked.FileInPath('cp3_llbb/Framework/data/ScaleFactors/Muon_TightRelIso_TightID_iso.json'),
+
+                hww_wp = cms.untracked.FileInPath('cp3_llbb/Framework/data/ScaleFactors/Muons_HWW_TnP_MediumIDandISO_Data_MC_25ns_PTvsETA_binBig_id_iso.json'),
                 )
             )
         )
